Added the ability to know where pacman is
Added a level object with a board data structure, this data structure
keeps record of the level in coordinates, it gives wether a coordinate
is valid or not, valid means Pacman can access to it. Pacman now also
has the ability to know in which place in the board is, so it can go to
any other part of it easily.
This change was made to be able to create a reaction agent.